use proposal title as the title of the proposal page
move function from useGovernanceProposal file so they could be used on server
