+++ src/features/governance/components/ProposalVoteChart.tsx
@@ -2,7 +2,6 @@ import { useMemo } from 'react';
 import { SpinnerWithLabel } from 'src/components/animation/Spinner';
 import { ColoredChartDataItem, StackedBarChart } from 'src/components/charts/StackedBarChart';
 import { Amount, formatNumberString } from 'src/components/numbers/Amount';
-import { MergedProposalData } from 'src/features/governance/hooks/useGovernanceProposals';
 import {
   useIsProposalPassing,
   useProposalQuorum,
@@ -23,6 +22,7 @@ import { fromWei } from 'src/utils/amount';
 import { bigIntSum, percent } from 'src/utils/math';
 import { objKeys } from 'src/utils/objects';
 import { toTitleCase } from 'src/utils/strings';
+import { MergedProposalData } from '../governanceData';
 
 export function PastProposalVoteChart({ id, title = 'Result' }: { id: number; title?: string }) {
   const { isLoading, votes } = useHistoricalProposalVoteTotals(id);
